refactor: restructure tab group components
