+++ src/app/components/TabGroup/index.ts
@@ -1,0 +1,3 @@
+export * from './TabGroup';
+export * from './Tab';
+export * from './TabPanel';
