+++ src/app/components/TabGroup/tabgroup.css
@@ -1,0 +1,103 @@
+.tabs {
+  width: 20em;
+  font-family: inherit;
+}
+
+[role='tablist'] {
+  margin: 0 0 -0.1em;
+  overflow: visible;
+}
+
+[role='tab'] {
+  position: relative;
+  margin: 0;
+  padding: 0.3em 0.5em 0.4em;
+  border: 1px solid gainsboro;
+  border-radius: 0.2em 0.2em 0 0;
+  box-shadow: 0 0 0.2em gainsboro;
+  overflow: visible;
+  font-family: inherit;
+  font-size: inherit;
+  background: #eceef2;
+}
+
+[role='tab']:hover::before,
+[role='tab']:focus::before,
+[role='tab'][aria-selected='true']::before {
+  position: absolute;
+  bottom: 100%;
+  right: -1px;
+  left: -1px;
+  border-radius: 0.2em 0.2em 0 0;
+  border-top: 3px solid #f05305;
+  content: '';
+}
+
+[role='tab'][aria-selected='true'] {
+  border-radius: 0;
+  background: #fbfcfe;
+  outline: 0;
+}
+
+[role='tab'][aria-selected='true']:not(:focus):not(:hover)::before {
+  border-top: 5px solid #29aed6;
+}
+
+[role='tab'][aria-selected='true']::after {
+  position: absolute;
+  z-index: 3;
+  bottom: -1px;
+  right: 0;
+  left: 0;
+  height: 0.3em;
+  background: #fbfcfe;
+  box-shadow: none;
+  content: '';
+}
+
+[role='tab']:hover,
+[role='tab']:focus,
+[role='tab']:active {
+  outline: 0;
+  border-radius: 0;
+  color: inherit;
+}
+
+[role='tab']:hover::before,
+[role='tab']:focus::before {
+  border-color: #f05305;
+}
+
+[role='tabpanel'] {
+  position: relative;
+  z-index: 2;
+  padding: 0.5em 0.5em 0.7em;
+  border: 1px solid gainsboro;
+  border-radius: 0 0.2em 0.2em 0.2em;
+  box-shadow: 0 0 0.2em gainsboro;
+  background: #fbfcfe;
+}
+
+[role='tabpanel']:focus {
+  border-color: #f05305;
+  box-shadow: 0 0 0.2em #f05305;
+  outline: 0;
+}
+
+[role='tabpanel']:focus::after {
+  position: absolute;
+  bottom: 0;
+  right: -1px;
+  left: -1px;
+  border-bottom: 3px solid #f05305;
+  border-radius: 0 0 0.2em 0.2em;
+  content: '';
+}
+
+[role='tabpanel'] p {
+  margin: 0;
+}
+
+[role='tabpanel'] * + p {
+  margin-top: 1em;
+}
